refactor: #63 utils 로 패키지 구조 변경
